backend/requirements.txt
fastapi==0.109.1
uvicorn[standard]==0.27.0
sqlalchemy==2.0.25
asyncpg==0.29.0
pgvector==0.2.4
alembic==1.13.1
scikit-learn==1.4.0
numpy==1.26.3
pandas==2.1.4
sentence-transformers==2.3.1
scipy==1.12.0
python-jose[cryptography]==3.4.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2
redis==5.0.1
pydantic==2.5.3
pydantic-settings==2.1.0
boto3==1.34.14
httpx==0.26.0
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
python-dotenv==1.0.0
python-multipart==0.0.22
aiosqlite==0.19.0
slowapi==0.1.9
joblib
